requirements.txt
Django~=5.2.4
djangorestframework~=3.16.0
drf-yasg~=1.21.10
pandas~=2.3.1
psycopg2-binary>=2.9
python-decouple>=3.8
uvicorn[standard]
